Start of Typing Mechanic
Made a first iteration of the Typing mechanic, which still needs a few tweaks fixed.
You can now type one of the three groups of words, which can be used to select a dialogue option. When you have typed all words within a group, that will be the corresponding dialogue option chosen.