Chronological Typing Mechanic
Instead of typing random words from a dialogue option to choose it, you now have to type words chronologically for the individual dialogue options.

Under each "DIalogueOption#" gameobject in the hierarchy is a "Words Added Per Mathematical Division_Max5" serializefield. It is curently set to 3, which means every third word of a dialogue option string counts for a new word having to be typed by the user.