Dialogue Animations + UI
In DialogueTesting scene the dialogue now reacts to tags in the ink file and changes the name, portrait and layout of the UI to fit the tags. The UI has also been updated to look more like a facebook chat. Still needs a way to see previously said dialogues above the current one "a chat history".