Ignore Implementation + Ending + Fixes
In act 3 if the timer runs out the chat will automatically change to the ignored ink file. Checks when both act 3 chats are done and then fades to an end screen based on the choices in act 3. Still needs to update ink files to make the act work properly + fix fade canvas to scale properly and show an interactable button to end the game as it is currently not raycastable but if it is enabled it stops the game from working as it overlaps the other canvas.